+++ src/react/apis/string/delete-string-api.ts
@@ -1,0 +1,11 @@
+import makeUseApiMutation from '../ApiMutation'
+import { stringController } from '../../constants/urls'
+import { IAxiosRequestConfig } from '../AxiosApi'
+
+export const useDeleteStringMutation = makeUseApiMutation<IAxiosRequestConfig<StringId>, boolean, StringId>(
+  'DELETE',
+  (id: StringId) => {
+    return `${stringController}/${id}`
+  })
+
+type StringId = string