+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [2022-07-12]
+### Added
+- delete string from backend
+
 ## [2022-07-11]
 ### Added
 - create/fetch strings from backend